Add files via upload
Made the CreateTables class - there were some problems with constraints (might be JavaDB related?  it will implement enough constraints for the file to work, so it's only a few of the later ones that don't work)
Updated the GamaManager to read from the database
Updated the DBManager to reflect the created tables, and to add code to read questions and answers from the database.